requirements.txt
pdf2image==1.17.0
python-pptx==1.0.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.162.0
pillow==11.1.0